Support Cache (#7)
Support a simple in-memory and persistent cache system. It keeps an
in-memory staging cache, which gets updated through the update method
and can be persisted through the commit method. Cache lookup is always
performed on the persistent cache only.

The cache will be applied on the results of both LLMChain and web search
tool, so that the result is always stable after commit() method is
called