Add a simple end-to-end test case (#139)
Add a simple end-to-end test case. 
We already setup a OPENAI_API_KEY for this project and the test should
be able to pass.

Note: Secrets set in a repository are not passed to workflows that are
triggered by a pull request from a fork. This is a security feature to
prevent exposure of secrets to potentially untrusted code.
Thus, this test suite will only be executed when a PR is merged.
Developers can also execute it if OPENAI_API_KEY env variable is setup.