Implement Auto-Retry Mechanism for Python Code Generation (#159)
Introduce auto-retry framework for python code generation. The framework
can be used on:
* df.ai.plot
* df.ai.verify
* UDF generation.

![output](https://github.com/pyspark-ai/pyspark-ai/assets/1097932/1e2dc7ac-c90b-4827-9cab-6b6c88ef7dc1)


This PR starts with `df.ai.plot`. After changes, the plotting will be
like:
```
[CODE GEN 1]
INFO: Getting the following error: 
[UNRESOLVED_COLUMN.WITH_SUGGESTION] A column or function parameter with name `start_date` cannot be resolved. Did you mean one of the following? [`start_lat`, `start_lon`, `end_lat`, `end_lon`, `airline`, `airport1`, `airport2`, `cnt`].
INFO: Retrying with 3 retries left
INFO: I apologize for the confusion. It seems there was a misunderstanding in the requirements. If the DataFrame doesn't have a 'start_date' column, we can't filter for flights in February 2011. 
[CODE GEN 2]
...
```
Also, this PR:
1. improves the prompt of plotting
2. changes the return type of `df.ai.plot` from None to str, so that it
is easier for users to get the generated code, and easier to test
plotting.